refactor: merge Mutable and normal classes.
