+++ src/main/java/cc/oakk/ladder/model/Ladder.java
@@ -1,5 +1,6 @@
 package cc.oakk.ladder.model;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.function.Function;
@@ -7,24 +8,22 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 import cc.oakk.ladder.model.line.Line;
-import cc.oakk.ladder.model.line.MutableLine;
-import cc.oakk.ladder.model.size.LadderSize;
+import cc.oakk.ladder.model.size.LadderHeight;
+import cc.oakk.ladder.model.size.LadderWidth;
 
 public class Ladder {
-    private final LadderSize size;
-    private final List<MutableLine> lines;
+    private final List<Line> lines;
 
     public Ladder(int width, int height) {
-        this.size = new LadderSize(width, height);
-        this.lines = Collections.unmodifiableList(IntStream.range(0, size.height().get()).boxed()
-                            .map(dummy -> new MutableLine(size.width().get()))
+        new LadderWidth(width);
+        new LadderHeight(height);
+        this.lines = Collections.unmodifiableList(IntStream.range(0, height).boxed()
+                            .map(dummy -> new Line(width))
                             .collect(Collectors.toList()));
     }
 
     public List<Line> getLines() {
-        return lines.stream()
-                .map(line -> (Line) line)
-                .collect(Collectors.toList());
+        return lines;
     }
 
     public Ladder initLines(Function<Integer, int[]> eval) {
