--- src/main/java/cc/oakk/ladder/model/line/MutableConnection.java
@@ -1,12 +1,0 @@
-package cc.oakk.ladder.model.line;
-
-public class MutableConnection extends Connection {
-    public MutableConnection(boolean connected) {
-        super(connected);
-    }
-
-    public MutableConnection set(boolean connected) {
-        this.connected = connected;
-        return this;
-    }
-}
